Implement best-first search. Can still use depth-first by changing the queue to LIFO.
This fixes the performance regression in test case eil33() vs. the iterative solver.
So it may be possible to discard the iterative code soon; it's clumsy and probably would
never be the right choice in the real world.

However, best-first search hurts performance in the hard problem instances like
eil33_moreVehicles(); it never really finds a feasible solution. A heuristic starting
point would certainly help.